PathCreatorComponent and TokenCounterComponent were returned to piplinetree.py
